\name{control.seq}
\alias{control.seq}
\title{control sequence for CTCF peaks}
\description{
  2500 control sequences for CTCF peaks in mouse (mm9) myotube.
}

\details{
  Control sequences are selected from 200bp window flanking the CTCF peaks with random offset.
}

\references{
  Unpublished
}

\examples{
  data(control.seq)
  control.seq
}

\keyword{datasets}
