\name{ctcf.seq}
\alias{ctcf.seq}
\title{sequence for CTCF peaks}
\description{
  a random subset of 2500 sequences for CTCF peaks in mouse (mm9) myotube.
}

\details{
  Sequences are extracted from 200bp window centering at peak summits.
}

\references{
  Unpublished
}

\examples{
  data(ctcf.seq)
  ctcf.seq
}

\keyword{datasets}
