\name{findMotifFasta}
\alias{findMotifFasta}
\title{Wrapper function for findMotif using fasta input}
\description{
  Perform motif search on two input fasta files. This is a wrapper
  function for findMotif
}
\usage{
findMotifFasta(fg.file, bg.file, ...)
}
\arguments{
  \item{fg.file}{character; foreground fasta filename}
  \item{bg.file}{character; background fasta filename}
  \item{...}{Other parameters passed to findMotif}
}
