\name{getSequence}
\alias{getSequence}
\title{Fetch genomic sequences from GRanges Object}
\description{
  The function fetch genomic sequences given the coordinates and strand informaton
}
\usage{
getSequence(gr, genome)
}

\arguments{
  \item{gr}{GRanges object; coordinates of sequences.}
  \item{genome}{BSgenome object}
}


